Move Create room button to sidebar
